+++ renpy/config.py
@@ -99,6 +99,15 @@ image_cache_size = None
 # The size of the image cache, in megabytes.
 image_cache_size_mb = 400
 
+# The number of threads to use for parallel image decoding during preloading.
+# Set to 0 for automatic (based on CPU count, capped as below and always leaving
+# at least 2 free) or 1 to disable parallel decoding.
+preload_threads = 0
+
+# The maximum number of threads to use for preloading when the above is
+# set to automatic.
+preload_thread_autocap = 4
+
 # The number of statements we will analyze when doing predictive
 # loading. Please note that this is a total number of statements in a
 # BFS along all paths, rather than the depth along any particular
@@ -1627,7 +1636,7 @@ extend_like_characters: set[str] = { "extend" }
 A set of character names that will be treated like the "extend" character for the purpose of dialogue export.
 """
 
-tlid_only_considers_say: bool = False
+tlid_only_considers_say: bool = True
 """
 If True, only say statements will be assigned translation ids.
 """
@@ -1635,6 +1644,12 @@ If True, only say statements will be assigned translation ids.
 safe_text: bool = False
 """
 If True, invalid text is displayed. If False, it's ignored.
+"""
+
+special_directory_map: dict[str, list[str]] = { 'images' : [ 'images' ], 'audio' : [ 'audio' ], 'fonts' : [ 'fonts' ] }
+"""
+This maps the special directory names ('images', 'audio', 'fonts') to a list of directories that will
+be searched for that kind of file.
 """