+++ renpy/display/behavior.py
@@ -2240,6 +2240,8 @@ class Bar(renpy.display.displayable.Displayable):
     """
 
     _store_transform_event = True
+    extra_changed = None
+    action = None
 
     @property
     def _draggable(self):
@@ -2275,6 +2277,7 @@ class Bar(renpy.display.displayable.Displayable):
         hovered=None,
         unhovered=None,
         released=None,
+        action=None,
         **properties,
     ):
         self.value = None
@@ -2286,6 +2289,7 @@ class Bar(renpy.display.displayable.Displayable):
 
                 self.value = value
                 adjustment = value.get_adjustment()
+                self.extra_changed = changed
 
                 if renpy.game.interface is not None:
                     renpy.game.interface.timeout(0)
@@ -2330,6 +2334,7 @@ class Bar(renpy.display.displayable.Displayable):
         self.unhovered = unhovered
 
         self.released = released
+        self.action = action
 
     def per_interact(self):
         if self.value is not None:
@@ -2630,6 +2635,16 @@ class Bar(renpy.display.displayable.Displayable):
             rv = self.adjustment.change(value)
             if rv is not None:
                 return rv
+
+            if self.extra_changed is not None:
+                rv = run(self.extra_changed, value)
+                if rv is not None:
+                    return rv
+
+            if self.action is not None:
+                rv = run(self.action)
+                if rv is not None:
+                    return rv
 
         if ignore_event:
             raise renpy.display.core.IgnoreEvent()