+++ renpy/display/particle.py
@@ -262,9 +262,12 @@ class SpriteManager(renpy.display.displayable.Displayable):
         self.width = None
         self.height = None
 
-    _duplicatable = True
+    _duplicatable = False
 
     def _duplicate(self, args):
+        if not self._duplicatable:
+            return self
+
         return SpriteManager(
             update=self.update_function,
             event=self.event_function,
@@ -431,6 +434,7 @@ class Particles(renpy.display.displayable.Displayable, renpy.rollback.NoRollback
         super(Particles, self).__init__(**properties)
 
         self.sm = SpriteManager(update=self.update_callback, predict=self.predict_callback, animation=animation)
+        self.sm._duplicatable = True
 
         self.factory = factory
         self.particles = None
@@ -440,7 +444,9 @@ class Particles(renpy.display.displayable.Displayable, renpy.rollback.NoRollback
     _duplicatable = True
 
     def _duplicate(self, args):
-        return Particles(self.factory, animation=self.sm.animation, **self.properties)
+        rv = Particles(self.factory, animation=self.sm.animation, **self.properties)
+        rv._duplicatable = False
+        return rv
 
     def update_callback(self, st):
         particles = self.particles