+++ renpy/substitutions.py
@@ -34,7 +34,7 @@ import collections
 
 
 update_translations = "RENPY_UPDATE_TRANSLATIONS" in os.environ
-flags = frozenset("rstiqulc!")
+flags = frozenset("rstiqulcf!")
 formatter = string.Formatter()
 
 SIMPLE_NAME = re.compile(r"^[a-zA-Z_][a-zA-Z0-9_]*$")
@@ -287,6 +287,9 @@ def convert(value, conv, scope):
             value = interpolate(value, scope)
         except RecursionError:
             raise ValueError(f"Substitution {value!r} refers to itself in a loop.")
+
+    if "f" in conv and renpy.config.say_menu_text_filter is not None:
+        value = renpy.config.say_menu_text_filter(value)
 
     if "q" in conv:
         value = value.replace("{", "{{")